+++ pkg/table/column.go
@@ -39,49 +39,12 @@ import (
 	"github.com/pingcap/tidb/pkg/types"
 	"github.com/pingcap/tidb/pkg/util/chunk"
 	"github.com/pingcap/tidb/pkg/util/hack"
-	"github.com/pingcap/tidb/pkg/util/intest"
 	"github.com/pingcap/tidb/pkg/util/logutil"
 	"github.com/pingcap/tidb/pkg/util/timeutil"
 	"go.uber.org/zap"
 )
 
-// Column provides meta data describing a table column.
-type Column struct {
-	*model.ColumnInfo
-	// If this column is a generated column, the expression will be stored here.
-	GeneratedExpr *ClonableExprNode
-	// If this column has default expr value, this expression will be stored here.
-	DefaultExpr ast.ExprNode
-}
-
-// ClonableExprNode is a wrapper for ast.ExprNode.
-type ClonableExprNode struct {
-	ctor     func() ast.ExprNode
-	internal ast.ExprNode
-}
-
-// NewClonableExprNode creates a ClonableExprNode.
-func NewClonableExprNode(ctor func() ast.ExprNode, internal ast.ExprNode) *ClonableExprNode {
-	return &ClonableExprNode{
-		ctor:     ctor,
-		internal: internal,
-	}
-}
-
-// Clone makes a "copy" of internal ast.ExprNode by reconstructing it.
-func (n *ClonableExprNode) Clone() ast.ExprNode {
-	intest.AssertNotNil(n.ctor)
-	if n.ctor == nil {
-		return n.internal
-	}
-	return n.ctor()
-}
-
-// Internal returns the reference of the internal ast.ExprNode.
-// Note: only use this method when you are sure that the internal ast.ExprNode is not modified concurrently.
-func (n *ClonableExprNode) Internal() ast.ExprNode {
-	return n.internal
-}
+type Column model.Column
 
 // String implements fmt.Stringer interface.
 func (c *Column) String() string {
